untextured rc working

+++ src/init_raycast.c
@@ -6,7 +6,7 @@
 /*   By: kkaczoro <kkaczoro@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/06/06 10:55:55 by kkaczoro          #+#    #+#             */
-/*   Updated: 2023/06/06 14:50:12 by kkaczoro         ###   ########.fr       */
+/*   Updated: 2023/06/07 10:00:22 by kkaczoro         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -17,7 +17,7 @@ static void	init_step_and_side_dist(t_raycast *rc, t_data *d);
 
 void	init_raycast(t_raycast *rc, t_data *data, int x)
 {
-	rc->camera_x = 2 * x / WINDOW_WIDTH - 1;
+	rc->camera_x = 2 * x / (double)WINDOW_WIDTH - 1;
 	rc->ray_dir_x = data->direction_x + data->plane_x * rc->camera_x;
 	rc->ray_dir_y = data->direction_y + data->plane_y * rc->camera_x;
 	rc->map_x = (int)data->player_x;
